LayoutTests/fast/events/spatial-navigation/snav-offscreen-content-expected.txt

offscreen contentRight link

Down link
PASS document.activeElement.getAttribute("id") is "1"
